Fix some slide errors

+++ presentation/chapters/shared/code/unsafe/2.rs
@@ -1,5 +1,5 @@
 #[inline]
-fn split_at_mut(&mut self, mid: usize) -> (&mut [T], &mut [T]) {
+fn split_at_mut(&mut self, mid: isize) -> (&mut [T], &mut [T]) {
     let len = self.len();
     let ptr = self.as_mut_ptr();
 
@@ -7,6 +7,6 @@ fn split_at_mut(&mut self, mid: usize) -> (&mut [T], &mut [T]) {
         assert!(mid <= len);
 
         (from_raw_parts_mut(ptr, mid),
-         from_raw_parts_mut(ptr.offset(mid as isize), len - mid))
+         from_raw_parts_mut(ptr.offset(mid), len - mid))
     }
 }
